Adds gradient utils and matches assets card to account card color
